backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
python-dotenv
pypdf
chromadb
openai
google-generativeai
serpapi